Restructuring directory, added time column to access log view
