+++ src/main/java/com/uts/iotbay/controller/RegisterServlet.java
@@ -1,4 +1,4 @@
-package com.uts.iotbay;
+package com.uts.iotbay.controller;
 
 /*
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
@@ -6,11 +6,13 @@ package com.uts.iotbay;
  */
 
 import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.StringWriter;
+
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.RequestDispatcher;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
@@ -20,8 +22,7 @@ import java.sql.ResultSet;
  *
  * @author michaellunn
  */
-//@WebServlet(urlPatterns = {"/login"})
-public class LogoutServlet extends HttpServlet {
+public class RegisterServlet extends HttpServlet {
 
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
@@ -32,9 +33,7 @@ public class LogoutServlet extends HttpServlet {
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
-    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-    }
+    
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
     /**
@@ -48,7 +47,6 @@ public class LogoutServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-//        processRequest(request, response);
     }
 
     /**
@@ -66,49 +64,52 @@ public class LogoutServlet extends HttpServlet {
             Class.forName("com.mysql.cj.jdbc.Driver");
             Connection con = DriverManager.getConnection("jdbc:mysql://localhost:3306/iotbay", "root", "iotbay");
             
-            User user = (User) request.getSession().getAttribute("user");
+            String email = request.getParameter("email");
+            String password = request.getParameter("password");
+            String fname = request.getParameter("fname");
+            String surname = request.getParameter("surname");
+            String phone = request.getParameter("phone");
 
-            String email = user.getEmail();
-            request.getSession().invalidate();
-            
-            PreparedStatement findUser = con.prepareStatement("select * from users where user_email=?");
+            PreparedStatement createUser = con.prepareStatement("INSERT INTO Users(user_email, user_password, user_fname, user_surname, user_type, user_phone, user_active) VALUES (?, ?, ?, ?, \"C\",?,1)");
+            createUser.setString(1, email);
+            createUser.setString(2, password);
+            createUser.setString(3, fname);
+            createUser.setString(4, surname);
+            createUser.setString(5, phone);
+            createUser.executeUpdate();
+
+            PreparedStatement findUser = con.prepareStatement("SELECT user_id FROM Users WHERE user_email = ?");
             findUser.setString(1, email);
+
             ResultSet rs = findUser.executeQuery();
 
+            int userId = -1;
+
             if(rs.next()) {
-
-                int userId = rs.getInt("user_id");
-
-                String sqlInsert = "INSERT INTO AccessLogs (user_id, date_accessed, activity_type) VALUES (?, CURRENT_TIMESTAMP(),\"Successful Logout\")";
-                PreparedStatement logStatement = con.prepareStatement(sqlInsert);
+                userId = rs.getInt("user_id");
+                PreparedStatement logStatement = con.prepareStatement("INSERT INTO AccessLogs (user_id, date_accessed, activity_type) VALUES (?, CURRENT_TIMESTAMP(),\"Account Created\")");
                 logStatement.setInt(1, userId);
                 logStatement.executeUpdate();
-
-                RequestDispatcher rd = request.getRequestDispatcher("homedirect.jsp");
-                rd.forward(request, response);
             }
-            else {
-                RequestDispatcher rd = request.getRequestDispatcher("homedirect.jsp");
-                rd.forward(request, response);
-            }
+            
+            response.sendRedirect("login.jsp");
         }
         
         catch(Exception e){
-//            PrintWriter out = response.getWriter();
-//            out.println("<!DOCTYPE html>");
-//            out.println("<html>");
-//            out.println("<head>");
-//            out.println("<title>Servlet LoginServlet</title>");            
-//            out.println("</head>");
-//            out.println("<body>");
-//            StringWriter sw = new StringWriter();
-//            PrintWriter pw = new PrintWriter(sw);
-//            e.printStackTrace(pw);
-//            out.println(sw.toString());
-//            out.println("</body>");
-//            out.println("</html>");
+           PrintWriter out = response.getWriter();
+           out.println("<!DOCTYPE html>");
+           out.println("<html>");
+           out.println("<head>");
+           out.println("<title>Servlet LoginServlet</title>");            
+           out.println("</head>");
+           out.println("<body>");
+           StringWriter sw = new StringWriter();
+           PrintWriter pw = new PrintWriter(sw);
+           e.printStackTrace(pw);
+           out.println(sw.toString());
+           out.println("</body>");
+           out.println("</html>");
         }
-//        
     }
 
     /**
