+++ src/main/webapp/delete.html
@@ -1,0 +1,97 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<link rel="stylesheet" href="styles.css">
+	<link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Space+Mono">
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover">
+	<title>IoTBay</title>
+
+</head>
+	
+	<body>
+		<div class="backdrop">
+			<div class="button-wrapper"></div>
+			<div class="buttons">
+				<ul class="menu-bar">
+					<li><button class="btn" tabindex="-1"> <a class="button-text" href="homedirect.jsp">Home</a></button></li>
+					<li><button class="btn" tabindex="-1"> <a class="button-text" href="about.html">About</a></button></li>
+					<li><button class="btn" tabindex="-1"> <a class="button-text" href="products">Products</a></button></li>
+					<li><button class="btn" tabindex="-1"> <a class="button-text" href="contact.html">Contact</a></button></li>
+					<li><button id="mng-acc-btn" tabindex="-1"><a href="manageaccount.jsp"><img id="mng-acc-pic" src="./assets/account.png" alt="manage account button"></a></button></li>
+                    <li><button id="cart-btn" tabindex="-1"><a href="manageaccount.jsp"><img id="cart-pic" src="./assets/shopping-cart.png" alt="manage account button"></a></button></li>
+				</ul>
+			</div>
+			<br>
+			<div class="logo">
+				<img src="./assets/logo.png" id="logo" alt="logo">
+				<!-- Logo created using DreamStudio by stability.ai -->
+				<br>
+			</div>
+			<div>
+				<h1 class="heading-text">Access Logs</h1>
+			</div>
+            <br>
+                <h2 class="subheading-text">All logs for:</h2>
+                
+                <h2 class="email-text">m.lunn54@gmail.com</h2>
+                <br>
+                <div class="date-range">
+                    <form action="accesslogs" class="date-range-form" method="post">
+                        <label class="date-range-form-label" for="from">From:</label>
+                        <input class="date-range-form-input" type="date" id="from" name="from" value="null">
+                        <label class="date-range-form-label" for="to">To:</label>
+                        <input class="date-range-form-input" type="date" id="to" name="to" value="null">
+                        <input class="date-range-form-input" id="date-range-submit" type="submit" value="Filter">
+                    </form>
+                </div>
+                <br>
+                <div class="access-logs-container">
+                    <table class="access-log-table">
+                        <tr>
+<th class="access-logs-header">Date</th>
+<th></th>
+<th class="access-logs-header">Time</th>
+<th></th>
+<th class="access-logs-header">Activity</th>
+</tr>
+<tr>
+<td class="access-log">2024-05-12</td>
+<td class="access-log">|</td>
+<td class="access-log">02:14:18</td>
+<td class="access-log">|</td>
+<td class="access-log">Successful Login</td>
+</tr>
+<tr>
+<td class="access-log">2024-05-12</td>
+<td class="access-log">|</td>
+<td class="access-log">02:14:18</td>
+<td class="access-log">|</td>
+<td class="access-log">Failed Login</td>
+</tr>
+<tr>
+<td class="access-log">2024-05-12</td>
+<td class="access-log">|</td>
+<td class="access-log">02:14:18</td>
+<td class="access-log">|</td>
+<td class="access-log">Successful Login</td>
+</tr>
+
+
+                    </table>
+                </div>
+                <br><br><br><br>
+
+	    </div>
+        <div class="footer">
+            <br>
+            <ul class="footer-menu">
+                <li><button class="footer-button" tabindex="-1"><a class="footer-button-text" href="homedirect.jsp">Home</a></button></li>
+                <li><button class="footer-button" tabindex="-1"><a class="footer-button-text" href="about.html">About</a></button></li>
+                <li><button class="footer-button" tabindex="-1"><a class="footer-button-text" href="products">Products</a></button></li>
+                <li><button class="footer-button" tabindex="-1"><a class="footer-button-text" href="contact.html">Contact</a></button></li>
+            </ul>
+            <p class="bottom-text">By Groot | University of Technology | Autumn 2024</p>
+        </div>	
+	</body>
+</html>
